Extract SignatureHelpProvider class
